requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
geojson==2.5.0
idna==3.4
psycopg2-binary==2.9.6
pyowm==3.3.0
PySocks==1.7.1
requests==2.31.0
urllib3==2.0.3
